sample issue solved in tests
